Updating layouts for sidebar/work area
Using a 10 x 1 grid now, to allow sidebars to be a little more comfortable.
Now ratios are 3 / 7 or 7 / 3, which feels much much better.

Also changing the naming conventions, in an attempt to make it easier for other
people to understand. Now each area has:
  - a number assigned to it
  - areas that will be filled by the window have the number inside square
    brackets [1], and areas that don't are just represented by a number

For instance, a sidebar of size 3 with an empty area on the side of size 7 will
be tagged as:

  [ 3 ] - 7

Which means the window occupies 3 out of 10.

All layouts take the full vertical space.